Fix broken import paths
I think `82c500c` when it was merged in
`7710e1fcefec1d909ec91ef81c3a830fb5276f64` broke these paths.
Incidentally, that merge was not rebased before it was merged :'(.

I have another commit that updates the canary used, in the latest canary
the css packages have been split out and there are some fun looking
"BaseClass not a Constructor" type errors.

So that is something to look forward to. For now, this commit gets the
`npm run dev` command up and running again.